Provide `ClientServer` implementation for pyodide
The existing `ClientServer` implementation cannot work for the pyodide
runtime as it relies on threads. Instead the pyodide implementation
provides a custom `Transport` that maps `_send_data` calls from the
source directly to `_procedure_handler` calls on the destination.

This appears to be sufficient to enable the existing test suite to
"just work" under pyodide.
